requirements.txt
streamlit
pandas
easyocr
opencv-python-headless
numpy
Pillow
openpyxl
